+++ css-inline/src/html/mod.rs
@@ -9,6 +9,7 @@ mod serializer;
 
 pub(crate) use self::selectors::Specificity;
 pub(crate) use document::Document;
+pub(crate) use node::{NodeData, NodeId};
 pub(crate) use parser::InliningMode;
 use smallvec::SmallVec;